pyproject.toml
[project]
name = "async-mega.py"
authors = [{name = "NTFSvolume", email = "NTFSvolume@proton.me"}]
dependencies = [
    "aiohttp>=3.11.18",
    "aiolimiter>=1.2.1",
    "pycryptodome==3.19.0"
]
description = "Python library for the Mega.nz and Transfer.it API"
license = "Apache-2.0"
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">=3.11"
version = "1.11.0"

[project.optional-dependencies]
cli = [
    "async-mega-py[default]",
    "typer-slim>=0.21.1"
]
default = [
    "python-dotenv>=1.2.1",
    "rich>=14.0.0"
]

[project.urls]
Homepage = "https://github.com/NTFSvolume/mega.py"

[tool.coverage.run]
omit = [
    "**/__init__.py",
    ".tox/*",
    "dist/*",
    "tests/*"
]

[tool.pytest.ini_options]
addopts = "-s -v"
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"
minversion = "8.3"
norecursedirs = [".git"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "COM8",  # flake8-commas linter
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "FA102",  # future annotations
    "FURB188",  # slice-to-remove-prefix-or-suffix
    "I",  # isort
    "N",  # PEP8 naming conventions
    "PLR1716",  # boolean chained comparison
    "Q",  # flake8-quotes
    "RUF",  # RUF specific fixes
    "T20",  # flake8-print
    "TC",  # flake8-type-checking
    "TID",  # flake8-tidy-imports
    "UP",  # pyupgrade
    "W"  # pycodestyle warnings
]
extend-safe-fixes = [
    "TC"  # move import from and to TYPE_CHECKING blocks
]
ignore = [
    "COM812",  # missing-trailing-comma
    "E501",  # suppress line-too-long, let formatter decide
    "N806"  # uppercase variables in functions
]
unfixable = [
    "ERA"  # do not autoremove commented out code
]

[tool.ruff.lint.per-file-ignores]
"**/{tests,docs}/*" = [
    "E402",  # imports not at the top of the file.
    "T20"  # flake8-print
]
"__init__.py" = ["E402"]  # imports not at the top of the file.

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.tomlsort]
all = true
ignore_case = true
in_place = true
sort_first = ["name", "project", "select", "tool"]
spaces_before_inline_comment = 2
spaces_indent_inline_array = 4

[tool.uv.build-backend]
module-name = "mega"

[build-system]
build-backend = "uv_build"
requires = ["uv_build>=0.8.17,<0.9.0"]

[dependency-groups]
dev = [
    "coverage[toml]>=7.7.0",
    "pre-commit>=4.5.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "pytest>=8.3.5",
    "requests-mock>=1.12.1",
    "rich>=14.0.0",
    "ruff>=0.13.1",
    "typing-extensions>=4.15.0"
]
